[#3889] Log outgoing requests, not log entries
