+++ src/openforms/logging/templates/logging/events/outgoing_request_log_details_view_admin.txt
@@ -1,4 +1,4 @@
 {% load i18n %}
 {% blocktrans trimmed with lead=log.fmt_lead user=log.fmt_user timelinelog=log.pk %}
-   {{ lead }}: User {{ user }} viewed timelinelog {{ timelinelog }} in the admin
+   {{ lead }}: User {{ user }} viewed outgoing request log {{ timelinelog }} in the admin
 {% endblocktrans %}
